got the images to display, but they are shit quality sadly due to nature of the links this is how it is
inside the generateArtistHTML  functions holds the ways to do it, direct link doesn't work and neither does shared link so we are forced to do it via thumbnail. but I dont think its important or worthwhile to change all of the google drive links for better image quality, not worth our time.